Replace binary cache file with JSON ones.
As pickle is not considered secure I decided to have a look at other
options and JSON turns out to be just as quick (or actually even quicker)
than binary file so I'll stick with that for now.

Of course it seems kinda stupid to parse (Strict)YAML and dump it to
JSON for speed but I'm not yet sure whether I want to:
- Drop type checking.
- Edit JSON while modifying data.

I'll need to take a look at plain YAML and/or JSON with validation but
for now I'm sticking with this.